tests/wpt/web-platform-tests/bluetooth/server/getPrimaryServices/blocklisted-services-with-uuid.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Request for services. Does not return blocklisted service.';
const expected = new DOMException(
    'Origin is not allowed to access the service. Tip: Add the service ' +
    'UUID to \'optionalServices\' in requestDevice() options. ' +
    'https://goo.gl/HxfxSQ', 'SecurityError');

bluetooth_test(() => getConnectedHIDDevice({
  filters: [{services: ['device_information']}],
  optionalServices: ['human_interface_device']
})
    .then(({device}) => assert_promise_rejects_with_message(
        device.gatt.getPrimaryServices('human_interface_device'),
        expected)),
    test_desc);
</script>
